Change numbering scheme to 3xx, use %autorelease
The new numbering scheme is 3%fedora, i.e. the 300 + Fedora release
number is used as Version. The version will need to be updated once
when a new branch diverges. %autorelease is used in the Release field.

We get conflicts with pretty much _every_ pull request made to this
package. This is just wasting the time of contributors. Convert to
%autorelease to avoid this. The numbering for this package already
changed in the past and while changing it again is not ideal, it
doesn't matter that much, as the upgrade path will be preserved. The
benefits of not having to adjust the release field and solve changelog
conflicts in each pull request are more important.