%_fix_broken_configure_for_lto: preserve line continuations
This is an old patch from jjames, extracted from
https://src.fedoraproject.org/rpms/redhat-rpm-config/pull-request/116

From the comments there:

The final sed expression of %_fix_broken_configure_for_lto replaces the
rest of the line after the assignment operator, thereby discarding any
line continuations that may happen to be on that line. This causes
https://bugzilla.redhat.com/show_bug.cgi?id=1874144, which I have now
seen affect multiple packages. This patch does the same replacements,
but preserves line continuations.