Add riscv64 config (thanks David Abdurachmanov)
For the most part, we can just follow what other architectures
are doing. Java, Node.js and GAP are all available on riscv64.

Frame pointers should be enabled but the architecture, just
like s390x, doesn't support packed relocations at the moment,
so the relevant comment is updated accordingly.

The biggest difference compared to other architectures is that
we disable LTO by default. For gcc, that's just a temporary
measure to cope with the fact that it slows down builds
significantly, and we will likely reconsider it as faster
hardware shows up; for clang, LTO requires ld.gold, which is
not available on riscv64 and currently in the process of being
deprecated in Fedora.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>